Sensorbox.Pifile: upgrade for current kernel
There is a newer kernel available on the mirrors than in our image. This
results in a different kernel and kernel-headers version and therefore
one can not load the wireguard kernel module. This fixes this issue by
just upgrading all the software - the kernel included.

A manually configured wireguard VPN does now work.